Remove vmx indirection
vm respects and follows mgetty's clever locking protocol to avoid
clashing on the port's use.

   When mgetty is started, it first checks if a valid lock file held by
another process exists.  If it does, this means that the port is in use,
and mgetty will wait until the lock file goes away.  Invalid lock files,
e.g.  for nonexistent processes ("stale" locks), are ignored.

   Once the port is free, mgetty creates its own lockfile, initializes
the modem and removes its lock file again.  Then it waits for something
to happen on the port.  Note that it does not _read_ any characters, it
just checks if there are any available for reading by using 'poll()' or
'select()'.

   There are two possibilities once characters arrive, either a
different program (e.g.  'uucico') has started dialing out or a 'RING'
was sent by the modem.  In the first case, mgetty should leave the port
alone.  This is easy _if_ the program dialing out has created a valid
lock file: mgetty will find it, wait for it to go away and then exit
(which will cause 'init' to start a fresh 'mgetty' process, which will
then wait for the next call).