Fix state engine to work for tests and real
The original implementation would freeze when testing, because
it was expecting an event when none was available.

The commit that added the test "or not state.has_entry_actions()",
which caused the failure tried to address a configuration
with an empty rule, e.g.

living:
	;

It turns out that rather than adding that condition the correct approach
is to look for direct transitions (rather than non-event ones).

While at it, also modify tests to start from a clean slate in order to
fix some unrelated failures.